+++ packages/firebase_start_repository/lib/firebase_start_repository.dart
@@ -1,0 +1,4 @@
+///
+library firebase_start_repository;
+
+export "src/firebase_killswitch_repository.dart";