+++ packages/flutter_start/lib/flutter_start.dart
@@ -1,4 +1,16 @@
-export 'package:flutter_introduction/flutter_introduction.dart';
-export 'package:flutter_introduction_shared_preferences/flutter_introduction_shared_preferences.dart';
-export 'src/models/start_configuration.dart';
-export 'src/user_stories/flutter_start_userstory_navigator.dart';
+// ignore_for_file: directives_ordering
+
+/// Userstory
+library;
+
+export "package:start_repository_interface/start_repository_interface.dart";
+export "package:flutter_introduction/flutter_introduction.dart";
+
+export "src/flutter_start_navigator_userstory.dart";
+
+/// models
+export "src/models/flutter_start_options.dart";
+
+/// Screens
+export "src/screens/splash_screen.dart";
+export "src/screens/introductions_screen.dart";