+++ packages/flutter_start/lib/src/widgets/default_splash_screen.dart
@@ -1,6 +1,0 @@
-import 'package:flutter/material.dart';
-
-Text defaultSplashScreen(BuildContext context) => Text(
-      'iconinstagram',
-      style: Theme.of(context).textTheme.headlineLarge,
-    );