+++ packages/start_repository_interface/lib/src/interfaces/killswitch_repository_interface.dart
@@ -1,0 +1,6 @@
+/// Killswitch repository interface
+// ignore: one_member_abstracts
+abstract class KillswitchRepositoryInterface {
+  /// Checks if the killswitch is active for the current app.
+  Future<bool> isKillswitchActive();
+}