+++ packages/start_repository_interface/lib/src/local/local_killswitch_repository.dart
@@ -1,0 +1,7 @@
+import "package:start_repository_interface/src/interfaces/killswitch_repository_interface.dart";
+
+/// The local killswitch repository.
+class LocalKillswitchRepository implements KillswitchRepositoryInterface {
+  @override
+  Future<bool> isKillswitchActive() async => false;
+}